chore: Migrate package to @crystallized/controllers
BREAKING CHANGE: removed CrystallineElement
