+++ packages/controllers/rb2js.config.rb
