+++ packages/controllers/test/targets-controller.test.js
@@ -1,7 +1,7 @@
 import { fixture, assert, aTimeout, html as testhtml } from "@open-wc/testing"
 import { LitElement, html } from "lit"
 
-import { TargetsController } from "../dist/controllers"
+import { TargetsController } from "../dist"
 
 // Fixtures
 
@@ -30,6 +30,18 @@ class TestElement extends LitElement {
 }
 customElements.define("test-element", TestElement)
 
+class NestedTargetsComponent extends LitElement {
+  targets = new TargetsController(this)
+
+  static get targets() {
+    return {
+      "button": "button",
+      "buttons": ["button"]
+    }
+  }
+}
+customElements.define("targets-component", NestedTargetsComponent)
+
 // Tests
 
 describe("TargetsController", () => {
@@ -43,5 +55,24 @@ describe("TargetsController", () => {
 
     el.shadowRoot.querySelector("button").click()
     assert.equal(el.shadowRoot.querySelector("test-msg").textContent, "clicked!clicked!howdy")
+  })
+
+  it("supports nested targets", async() => {
+    const el = await fixture(testhtml`
+      <targets-component>
+        <button id="outer">Outer Button</button>
+        <targets-component>
+          <button id="inner">Inner Button</button>
+        </targets-component>
+      </targets-component>
+    `)
+
+    assert.equal(el._button, el.querySelector("#outer"))
+    assert.equal(el._buttons.length, 1)
+    assert.equal(el._buttons[0], el.querySelector("#outer"))
+
+    const nestedEl = el.querySelector("targets-component")
+    assert.equal(nestedEl._buttons.length, 1)
+    assert.equal(nestedEl._buttons[0], el.querySelector("#inner"))
   })
 })
